refactor: roles to give check
instead of using an object and setting the roleId's (key) value to the points if they have it, instead, it'll rely on the total points instead.

this is because, if the message sends, it's only when they reached a new requirement.

if the length is over 1 regardless, it wont send a message and just grant the roles (because they might've just left and rejoined)